move ecmdWriteTarget() and utils to separate file Signed-off-by: Kahn Evans <kahnevan@us.ibm.com>
